+++ src/lib.rs
@@ -4,3 +4,6 @@ mod alignment;
 mod alloc_free;
 mod alloc_result;
 mod memory;
+
+pub use alloc_result::AllocationError;
+pub use memory::Memory;